refactored AIAssistant and cleaned up process_query()
